Make consistent table column widths
Added explicit width values for first and second
table columns corresponding to "Recipe" and "Crates".
Please note that this is a broad CSS style for all tables
which will break any non TOC tables.

For a cleaner solution we would need to explicitly
add class attribute to the generated HTML.

resolves: https://github.com/brson/rust-cookbook/issues/42